Create accounts and commtents

+++ Python/django-form/boards/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST, require_GET, require_http_methods
-from .models import Board
-from .forms import BoardForm
+from django.contrib.auth.decorators import login_required
+from .models import Board, Comment
+from .forms import BoardForm, CommentForm
 
 
 @require_GET
@@ -11,6 +12,7 @@ def index(request):
     return render(request, 'boards/index.html', context)
 
 
+@login_required
 @require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == 'POST':
@@ -20,6 +22,8 @@ def create(request):
             # title = form.cleaned_data.get('title')  # cleaned_data : 값을 깔끔하게 다듬어서 꺼내는 요청
             # content = form.cleaned_data.get('content')
             # board = Board.objects.create(title=title, content=content)
+            board = form.save(commit=False)
+            board.user = request.user
             board = form.save()
             return redirect('boards:detail', board.pk)
     else:  # GET 요청 또는 유효성 검사를 충족하지 못한 경우
@@ -32,7 +36,13 @@ def create(request):
 @require_GET
 def detail(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    context = {'board': board}
+    comments = board.comment_set.order_by('-pk')
+    comment_form  = CommentForm()
+    context = {
+        'board': board,
+        'comment_form': comment_form,
+        'comments': comments,
+    }
     return render(request, 'boards/detail.html', context)
 
 
@@ -40,10 +50,13 @@ def detail(request, board_pk):
 @require_POST
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
+    if request.user != board.user:
+        return redirect('boards:detail', board_pk)
     board.delete()
     return redirect('boards:index')
 
 
+@login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
@@ -55,6 +68,8 @@ def update(request, board_pk):
             # board.title = form.cleaned_data.get('title')
             # board.content = form.cleaned_data.get('content')
             # board.save()
+            board = form.save(commit=False)
+            board.user = request.user
             board = form.save()  # instance=board -> Board 를 새로 생성하지않고 기존의 Board 를 업데이트
             return redirect('boards:detail', board.pk)
 
@@ -68,3 +83,32 @@ def update(request, board_pk):
         'board_pk': board_pk,
     }
     return render(request, 'boards/form.html', context)  # form 내에 데이터가 있는 상태로 보인다.
+
+
+@require_POST
+def comments_create(request, board_pk):
+    # 로그인하지 않은 사용자 차단
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+    # 댓글 작성 로직
+    comment_form = CommentForm(request.POST)  # model form 에 사용자 입력을 받는다.
+    # 유효성 검사
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False)
+    # 유저 정보 할당
+        comment.user = request.user
+        # 같은 말 comment.user_id = request.user.id
+    # board 정보 할당
+        comment.board_id = board_pk  # pk 로 할당
+    # 저장
+        comment.save()
+
+    return redirect('boards:detail', board_pk)
+
+
+def comments_delete(request, board_pk, comment_pk):
+    # 댓글 삭제 로직
+    comment = get_object_or_404(Comment, pk=comment_pk)
+    if comment.user == request.user:
+        comment.delete()
+    return redirect('boards:detail', board_pk)
